Update cis129_lab02_shortStory.py
replaced    print('Hello World!')   with

print('Once apon a time in a small town, there lived a old yet curious cat, named Whiskers.')
print('One day Whiskers was out for his daily stroll and noticed a rusty metal figure on the ground,')
print('as he aproched it he noticed it was an old key. He was unsure who or what it belonged to so he')
print('continued on his way and had seen a newspaper on the ground that stated "LOST TRESURE".')
print('Whiskers read the articale that stated a tresure chest had washed apon shore and was missing a key,')
print('Wihskers then took his chance and had taken the old rust key and requested to try and open it.')
print('Meow! It opened, he was then giving the contents inside which contained a gallon of milk.')
print('The End.')


Which results to a short story about Whiskers.